refactor(frontend): centralize environment config and add auto-validation
- Refactor gateway-client.ts to use centralized env.ts instead of direct process.env access
- Remove gatewayUrl instance variable, use env.GATEWAY_URL directly
- Add validateClientEnv() call in GatewayClient constructor for automatic environment validation
- Remove manual validation from page component
- Improve code consistency and developer experience with centralized config
